Move new scripts to the Scripts folder
