+++ Qubi/Assets/Scripts/CharacterBarVisuals.cs
@@ -1,0 +1,21 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class CharacterBarVisuals : MonoBehaviour
+{
+
+
+    private void Start()
+    {
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+
+
+
+    }
+}
