+++ Qubi/Assets/Scripts/DuplicateHorizontal.cs
